% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\docType{package}
\name{BLCAsubtyping-package}
\alias{BLCAsubtyping}
\alias{BLCAsubtyping-package}
\title{BLCAsubtyping - Predicting bladder cancer molecular subtypes}
\description{
This package provides the 'classify' function, that can be used to predict class for bladder 
cancer samples using several differente classification systems, from the sample's expression profile.
}
\note{
This is a contribution from the Tumor Identity Cards (CIT) program founded by the 
'Ligue Nationale Contre le Cancer' (France): \url{http://cit.ligue-cancer.net}. 
For any question please contact \url{CITR@ligue-cancer.net}
}
\seealso{
\link[BLCAsubtyping]{classify}
}
\author{
Aurelien de Reynies, Aurelie Kamoun
}
