% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\docType{data}
\name{cit}
\alias{cit}
\title{CIT BLCA cohort data}
\format{A list with two elements:
\describe{
    \item{expMat}{A data.frame that contains gene expression data derived from microarray, 
    with probes in the rows and samples in the columns.}
    \item{gpl}{ A data.frame that contains annotation for the probes, with different probes
    in the rows and identifiers in the columns.}
}}
\usage{
cit
}
\description{
Gene expression and annotation data derived from the CIT cohort of bladder cancer.
}
\references{
\url{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-1803/}
}
\keyword{datasets}
