% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bladderCancerSubtypingSystems.R
\name{classify}
\alias{classify}
\title{classify - Classify a BLCA sample from gene expression data}
\usage{
classify(expMat, gpl = NULL, symbol = "Gene.Symbol",
  classification.systems = c("Baylor", "UNC", "CIT", "Lund", "MDA",
  "TCGA"))
}
\arguments{
\item{expMat}{Matrix or dataframe of normalized expression data. 
Samples are in column and probesets are in row. Rownames and colnames are required.}

\item{gpl}{Dataframe with matching information between expMat rownames and HGNC gene symbols. 
The dataframe must have rownames corresponding to ExpMat rownames,
and at least one column with associated gene symbols}

\item{symbol}{Character string referring to the column name of gpl containing gene symbols}

\item{classification.systems}{A character vector with names of the classification systems to consider
among Baylor, UNC, CIT, Lund, MDA and TCGA.}
}
\value{
A dataframe of samples annotated according to each classification system.
}
\description{
This function assigns molecular subtypes to bladder samples using expression data, 
according to several classification systems.
}
\note{
This is a contribution from the Tumor Identity Cards (CIT) program founded by the 
'Ligue Nationale Contre le Cancer' (France): \url{http://cit.ligue-cancer.net}. 
For any question please contact \url{CITR@ligue-cancer.net}
}
\examples{
data(cit)

#-- Classify using all classification types
res <- classify(cit$expMat, cit$gpl, symbol = "Symbol")
head(res)

#-- Get TCGA classification
tcga_res <- res[, c("ID", "TCGA.subtype")]
head(tcga_res)

}
\author{
Aurelien de Reynies
}
\keyword{methods}
